pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ap-sachivalayam-copilot"
version = "0.1.0"
description = "AI Copilot for AP Village Secretariat Employees"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # Cache & Queue
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    # AI / LLM
    "anthropic>=0.39.0",
    "openai-whisper>=20231117",
    "sentence-transformers>=3.0.0",
    # WhatsApp
    "httpx>=0.27.0",
    # PDF Generation
    "weasyprint>=62.0",
    # Config & Security
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=43.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "rapidfuzz>=3.9.0",
    # Monitoring
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "locust>=2.31.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
